fix: notify removed participant before updating session
In a previous fix we incorrectly moved socket cleanup to after the final update was broadcast to session participants. This meant the removed participant recieved the update before being notified they'd been removed which caused a crash in the client

Instead, we still do clean-up and notification before final session update, but only if the socketId exists on the participant (ie. they haven't temporarily disconnected)